support fastpath in blkmq mode
