+++ Makefile.in
@@ -1,8 +1,10 @@
-px-objs = pxd.o dev.o iov_iter.o px_version.o io.o
+px-objs = pxd.o dev.o iov_iter.o px_version.o io.o kiolib.o pxd_fastpath.o pxd_bio_makereq.o pxd_bio_blkmq.o
 obj-m = px.o
 
 KBUILD_CPPFLAGS := -D__KERNEL__
 PXDEFINES := -D__PXKERNEL__
+#PXDEFINES += -D__PX_FASTPATH__ -D__PXD_BIO_MAKE_REQ__
+PXDEFINES += -D__PX_FASTPATH__ -D__PXD_BIO_BLKMQ__
 
 KVERSION=$(shell uname -r)
 
@@ -80,28 +82,6 @@ endif
 ifeq ($(shell echo "$(CHK_KVER)" | grep -Eq '.*\.el8.*\.x86_64'; echo $$?),0)
 PXDEFINES += -D__PX_BLKMQ__ -D__EL8__
 endif
-
-## fastpath specific checks
-ifeq ($(call verlater,${kmajor},${fp_major}),0)
-PXDEFINES += -D__PX_FASTPATH__
-px-objs += pxd_fastpath.o
-$(info Kernel fast path enabled, current kernel version $(CHK_KVER) need minimum $(FPATH_MINKVER))
-else
-ifeq ($(call versame,${kmajor},${fp_major}),0)
-ifeq ($(call versameorlater,${kminor},${fp_minor}),0)
-PXDEFINES += -D__PX_FASTPATH__
-px-objs += pxd_fastpath.o
-$(info Kernel fast path enabled, current kernel version $(CHK_KVER) need minimum $(FPATH_MINKVER))
-else
-px-objs += pxd_fastpath_stub.o
-$(warning Kernel fast path disabled, current kernel version $(CHK_KVER) need minimum $(FPATH_MINKVER))
-endif
-else
-px-objs += pxd_fastpath_stub.o
-$(warning Kernel fast path disabled, current kernel version $(CHK_KVER) need minimum $(FPATH_MINKVER))
-endif
-endif
-
 
 ifdef KERNELOTHER
 KERNELOTHEROPT=O=$(KERNELOTHER)
